Oracle abi taken from compilation. Moved project setup from dockerfile to package.json scripts
