+++ .circleci/config.yml
@@ -15,6 +15,7 @@ jobs:
       - checkout
       - run: git submodule update --init
       - run: yarn
+      - run: yarn prepare:contracts
       - run: yarn run lint
   ansible-lint:
     docker:
@@ -32,6 +33,7 @@ jobs:
       - checkout
       - run: git submodule update --init
       - run: yarn
+      - run: yarn prepare:contracts
       - run: yarn run test
   oracle-e2e:
     docker:
