+++ Dockerfile.e2e
@@ -5,6 +5,7 @@ COPY package.json .
 COPY oracle/package.json ./oracle/
 COPY oracle-e2e/package.json ./oracle-e2e/
 COPY ui/package.json ./ui/
+COPY ui/lib/web3-eth/index.js ./ui/lib/web3-eth/index.js
 COPY ui-e2e/package.json ./ui-e2e/
 COPY monitor/package.json ./monitor/
 COPY contracts/package.json ./contracts/
@@ -12,12 +13,10 @@ COPY ui/lib/web3-eth/index.js ./ui/lib/web3-eth/index.js
 
 COPY yarn.lock .
 RUN yarn install
+COPY ./contracts ./contracts
+RUN yarn install:deploy
 
-COPY contracts contracts
-RUN yarn workspace poa-parity-bridge-contracts run compile
-RUN cd contracts/deploy && rm -f package-lock.json && npm install
-
-COPY ui ui
-RUN yarn workspace ui run compile:contracts && yarn workspace ui run postinstall
+RUN mkdir -p oracle/abis monitor/abis ui/src/contracts
+RUN yarn prepare:contracts
 
 COPY . .
