+++ oracle/scripts/initialChecks.js
@@ -3,7 +3,7 @@ require('dotenv').config({
   path: path.join(__dirname, '../.env')
 })
 const Web3 = require('web3')
-const ERC677BridgeTokenABI = require('../abis/ERC677BridgeToken.abi')
+const ERC677BridgeTokenABI = require('../abis/ERC677BridgeToken').abi
 const { ERC_TYPES } = require('../src/utils/constants')
 
 async function initialChecks() {
